+++ packages/docs/pages/skills/library/cool.mdx
