+++ packages/docs/pages/skills/library/game.mdx
