+++ packages/docs/pages/skills/library/pay.mdx
