+++ packages/docs/pages/skills/library/register.mdx
