--- packages/docs/pages/templates/agent/index.mdx
@@ -1,84 +1,0 @@
-# Web3 Agent
-
-This is a simple yet powerful agent that will help users interact with web3 products.
-
-## Skills
-
-- [Check](/templates/agent/handlers/check): Check if a ENS domain is available
-- [Cool](/templates/agent/handlers/cool): Show cool domain alternatives
-- [Register](/templates/agent/handlers/register): Register a ENS domain
-- [Reset](/templates/agent/handlers/reset): Reset your conversation
-- [Renew](/templates/agent/handlers/renew): Renew your ENS domain
-- [Info](/templates/agent/handlers/info): Get info about a ENS domain
-- [Pay](/templates/agent/handlers/pay): Pay in usdc
-- [Token](/templates/agent/handlers/token): Fetch token price
-- [Game](/templates/agent/handlers/game): Play a game
-
-## Structure
-
-```
-agent/
-├── src/
-│   └── handler/            # Skills
-│       ├── check.ts
-│       ├── cool.ts
-│       ├── info.ts
-│       ├── register.ts
-│       ├── renew.ts
-│       ├── reset.ts
-│       ├── pay.ts
-│       ├── token.ts
-│       ├── game.ts        
-│   ├── index.ts            # Skill loading
-│   ├── prompt.ts           # Prompt
-├── package.json
-├── tsconfig.json
-└── .env
-```
-
-## Runner
-
-The runner function to start the agent and listen for incoming messages:
-
-```tsx [src/index.ts]
-// [!include ~/../../templates/agent/src/index.ts:skills]
-// [!include ~/../../templates/agent/src/index.ts:run]
-```
-
-## Prompt
-
-This is the prompt that the agent will use to respond to the user.
-
-```tsx [src/prompt.ts]
-// [!include ~/../../templates/agent/src/prompt.ts]
-```
-
-## Run the app
-
-Follow the steps below to run the app
-
-::::steps
-
-### Setup
-
-```bash [cmd]
-# Clone the repo
-git clone https://github.com/ephemeraHQ/message-kit
-# Go to the templates/agent folder
-cd templates/agent
-# Install the dependencies
-yarn install
-# run the app
-yarn dev
-```
-
-### Variables
-
-Set up these variables in your app
-
-```bash [.env]
-KEY= # the private key of the wallet
-OPENAI_API_KEY= # sk-proj-...
-```
-
-::::
