+++ packages/docs/vercel.json
@@ -7,7 +7,6 @@
       "destination": "/index.html"
     }
   ],
-  "rootDirectory": "packages/docs/",
   "env": {
     "VERCEL_FORCE_NO_CACHE": "1"
   }
